updated unit 6 and added unit 7

+++ unit06_true_false/task_10_bugs.py
@@ -1,5 +1,5 @@
-squareSide = int(input())
-bugSpeed = int(input())
+squareSide = float(input())
+bugSpeed = float(input())
 distance = squareSide
 seconds = 0
 if distance != bugSpeed:
